Refactor .h files to .hpp files
Refactors all left-over `.h` files and fixes cuSTL include orders
on the way.
