+++ src/libPMacc/examples/gameOfLife2D/main.cu
@@ -30,7 +30,7 @@
 #include "Simulation.hpp"
 
 #include <mpi.h>
-#include "communication/manager_common.h"
+#include "communication/manager_common.hpp"
 
 namespace po = boost::program_options;
