+++ src/libPMacc/include/cuSTL/container/DeviceBuffer.hpp
@@ -27,15 +27,17 @@
 #include "cuSTL/container/copier/D2DCopier.hpp"
 #include "cuSTL/container/assigner/DeviceMemAssigner.hpp"
 #include "cuSTL/container/CartBuffer.hpp"
-#include "cuSTL/container/allocator/tag.h"
+#include "cuSTL/container/allocator/tag.hpp"
 #include "cuSTL/container/copier/Memcopy.hpp"
 
 #include <boost/assert.hpp>
 #include <boost/move/move.hpp>
 #include <boost/utility/enable_if.hpp>
 #include <boost/type_traits/is_same.hpp>
+
 #include <exception>
 #include <sstream>
+
 
 namespace PMacc
 {
