+++ src/libPMacc/include/cuSTL/container/HostBuffer.hpp
@@ -26,15 +26,17 @@
 #include "cuSTL/container/copier/H2HCopier.hpp"
 #include "cuSTL/container/assigner/HostMemAssigner.hpp"
 #include "cuSTL/container/CartBuffer.hpp"
-#include "cuSTL/container/allocator/tag.h"
+#include "cuSTL/container/allocator/tag.hpp"
 #include "cuSTL/container/copier/Memcopy.hpp"
 
 #include <boost/assert.hpp>
 #include <boost/move/move.hpp>
 #include <boost/utility/enable_if.hpp>
 #include <boost/type_traits/is_same.hpp>
+
 #include <exception>
 #include <sstream>
+
 
 namespace PMacc
 {
