+++ src/libPMacc/include/cuSTL/container/allocator/HostMemAllocator.hpp
@@ -22,10 +22,13 @@
 
 #pragma once
 
-#include <stdint.h>
-#include "math/vector/Size_t.hpp"
+#include "tag.hpp"
 #include "cuSTL/cursor/BufferCursor.hpp"
-#include "tag.h"
+
+#include "math/vector/Size_t.hpp"
+
+#include <stdint.h>
+
 
 namespace PMacc
 {
