+++ src/libPMacc/include/lambda/CT/TerminalTL.hpp
@@ -22,13 +22,15 @@
 
 #pragma once
 
-#include "../ExprTypes.h"
-#include "../placeholder.h"
+#include "../ExprTypes.hpp"
+#include "../placeholder.hpp"
+#include "../Expression.hpp"
+#include "lambda/ProxyClass.hpp"
+
 #include <boost/mpl/vector.hpp>
 #include <boost/mpl/copy.hpp>
 #include <boost/mpl/front_inserter.hpp>
-#include "../Expression.hpp"
-#include <lambda/ProxyClass.hpp>
+
 
 namespace mpl = boost::mpl;
