+++ src/picongpu/include/plugins/PhaseSpace/PhaseSpace.hpp
@@ -23,7 +23,7 @@
 #include <mpi.h>
 
 #include "simulation_defines.hpp"
-#include "communication/manager_common.h"
+#include "communication/manager_common.hpp"
 #include "pluginSystem/INotify.hpp"
 #include "cuSTL/container/DeviceBuffer.hpp"
 #include "cuSTL/container/HostBuffer.hpp"
